change in comparisons in lines 81 and 132
The comparison scheme implemented in earlier code would result in an out of bounds error. It was not apparent  in the code because the dimensions of the camera being used were symmetric(128x128). However, in the case of a 240x180 camera's event data, the code would throw an error. 
Proposed change looks to rectify this discrepancy.